Generalized to N dimensions, Fixed rotation animation
N dimensions:

Z rotation:
cube[x][y] = cube[dim-1-y][x];

Rotation animation:

1. Increment the scene cubes' angles using the queued rotation. If angle is in endzone, rotate to 0 out the angle.
2. Render
3. Update the queued rotation angle. Dequeue if it is in endzone, otherwise update.

This ensures that when the queued rotation angle is in the endzone, the scene cubes will first be rendered to complete the rotation, and then the queued rotation will be dequeued.

If it is not in the endzone, the cubes will be rotated at the normal rate, and the queued rotation angle will be updated at the normal rate.