Made custom number spinner, Added Dimension ng controller methods
The glyphicons need to have position: absolute, so that they are positioned relative to their containing positioned ancestor.

The trick was to make sure that the nearest positioned ancestor is actually their parent (class inner-addon). Have to set inner-addon to have position: relative for the glyphicons to use it as the nearest positioned parent. If inner-addon is left to the default position: static, the glyphicons will treat the grandparent bootstrap col-sm-12 class as their nearest positioned parent.